move logic in a service
